fix(ci): split oversized diff viewers under repo-shape caps
Problem: Main CI failing repo-shape check — ChatDiffViewer (578→500) and
SplitDiffViewer (540→500) exceed max-lines cap. #941 addresses render_tests.rs;
this PR fixes the two new violations from morning's diff-viewer merges.

What changed:
- Extracted cell/row components → ChatDiffCells.tsx + SplitDiffCells.tsx
- Extracted gap-expansion helpers → chat-diff-rows.ts + split-diff-rows.ts
- Moved pure gap logic to lib/domain/files/gap-expansion.ts (frontend boundary compliance)
- ChatDiffViewer: 578 → 383 lines
- SplitDiffViewer: 540 → 299 lines

Split pattern matches #853/#854 (extract cohesive units to siblings, wire back).
No behavior changes — launch-critical UI.

How to verify:
- python3 scripts/check_max_lines.py (both files clean)
- python3 scripts/check_frontend_boundaries.py (passes)
- python3 scripts/report_frontend_structure.py --strict (no new violations)
- pnpm --filter @proliferate/desktop exec tsc --noEmit (typechecks)

Complements #941; together they turn main green.
